RemoteControl API documentation mostly complete Moved the API operations all into remoteControlApi.doxygen (for easier overview and TOC support) MainService move operation now supports float x/y movements to scale the speed Fixed landscapedescription redirection

+++ plugins/RemoteControl/src/MainService.hpp
@@ -39,12 +39,11 @@ class StelPropertyMgr;
 class StelScriptMgr;
 class StelSkyCultureMgr;
 
+//! @ingroup remoteControl
 //! Implements the main API services, including the \c status operation which can be repeatedly polled to find the current state of the main program,
 //! including time, view, location, StelAction and StelProperty state changes, movement, script status ...
 //!
-//! ## Service methods
-//! @copydetails getImpl()
-//! @copydetails postImpl()
+//! @see @ref rcMainService
 class MainService : public AbstractAPIService
 {
 	Q_OBJECT
@@ -57,98 +56,11 @@ public:
 	virtual void update(double deltaTime) Q_DECL_OVERRIDE;
 
 protected:
-	//! @brief Implements the GET method
-	//! @details
-	//! ### GET operations
-	//! #### status
-	//! Parameters: <tt>[actionId (Number)] [propId (Number)]</tt>\n
-	//! This operation can be polled every few moments to find out if some primary Stellarium state changed. It returns a JSON object with the following format:
-	//! @code{.js}
-	//! {
-	//!	//current location information, see StelLocation
-	//!	location : {
-	//!		name,
-	//!		role,
-	//!		planet,
-	//!		latitude,
-	//!		longitude,
-	//!		altitude,
-	//!		country,
-	//!		state,
-	//!		landscapeKey
-	//!	},
-	//!	//current time information
-	//!	time : {
-	//!		jday,		//current Julian day
-	//!		deltaT,		//current deltaT as determined by the current dT algorithm
-	//!		gmtShift,	//the timezone shift to GMT
-	//!		timeZone,	//the timezone name
-	//!		utc,		//the time in UTC time zone as ISO8601 time string
-	//!		local,		//the time in local time zone as ISO8601 time string
-	//!		isTimeNow,	//if true, the Stellarium time equals the current real-world time
-	//!		timerate	//the current time rate (in secs)
-	//!	},
-	//!	script : {
-	//!		scriptIsRunning,	//boolean, true if a script is running
-	//!		runningScriptId		//the ID of the running script
-	//!	},
-	//!	selectioninfo, //string that contains the information of the currently selected object, as returned by StelObject::getInfoString
-	//!	view : {
-	//!		fov,		//current FOV
-	//!		projection,	//current projection key
-	//!		projectionStr,	//current projection string in current language
-	//!		landscape,	//current landscape ID
-	//!		skyculture	//current skyculture ID
-	//!	},
-	//!
-	//!	//the following is only inserted if an actionId parameter was given
-	//!	//see below for more info
-	//!	actionChanges : {
-	//!		id, //currently valid action id, the interface should update its own id to this value
-	//!		changes : {
-	//!			//a list of boolean actions that changed since the actionId parameter
-	//!			<actionName> : <actionValue>
-	//!		}
-	//!	},
-	//!	//the following is only inserted if an propId parameter was given
-	//!	//see below for more info
-	//!	propertyChanges : {
-	//!		id, //currently valid prop id, the interface should update its own id to this value
-	//!		changes : {
-	//!			//a list of properties that changed since the propId parameter
-	//!			<propName> : <propValue>
-	//!		}
-	//!	}
-	//! }
-	//! @endcode
-	//!
-	//! The \c actionChanges and \c propertyChanges sections allow a remote interface to track boolean StelAction and/or StelProperty changes.
-	//! On the initial poll, you should pass -2 as \p propId and \p actionId. This indicates to the service that you want a full
-	//! list of properties/actions and their current values. When receiving the answer, you should set your local \p propId /\p actionId to the id
-	//! contained in \c actionChanges and \c propertyChanges, and re-send it with the next request as parameter again.
-	//! This allows the MainService to find out which changes must be sent to you (it maintains a queue of action/property changes internally, incrementing
-	//! the ID with each change), and you only have to process the differences instead of everything.
-	//! #### plugins
-	//! Returns the list of all known plugins, as a JSON object of format:
-	//! @code{.js}
-	//! {
-	//!	//list of known plugins, in format:
-	//!	<pluginName> : {
-	//!		loadAtStartup,	//if to load the plugin at startup
-	//!		loaded,		//if the plugin is currently loaded
-	//!		//corresponds to the StelPluginInfo of the plugin
-	//!		info : {
-	//!			authors,
-	//!			contact,
-	//!			description,
-	//!			displayedName,
-	//!			startByDefault,
-	//!			version
-	//!		}
-	//!	}
-	//! }
-	//! @endcode
+	//! @brief Implements the GET operations
+	//! @see @ref rcMainServiceGET
 	virtual void getImpl(const QByteArray& operation,const APIParameters &parameters, APIServiceResponse& response) Q_DECL_OVERRIDE;
+	//! @brief Implements the HTTP POST operations
+	//! @see @ref rcMainServicePOST
 	virtual void postImpl(const QByteArray &operation, const APIParameters &parameters, const QByteArray &data, APIServiceResponse &response) Q_DECL_OVERRIDE;
 
 private slots:
@@ -161,7 +73,7 @@ private slots:
 	bool focusObject(const QString& name);
 	void focusPosition(const Vec3d& pos);
 
-	void updateMovement(int x, int y, bool xUpdated, bool yUpdated);
+	void updateMovement(float x, float y, bool xUpdated, bool yUpdated);
 	void setFov(double fov);
 
 	void actionToggled(const QString& id, bool val);
@@ -178,7 +90,7 @@ private:
 	StelScriptMgr* scriptMgr;
 	StelSkyCultureMgr* skyCulMgr;
 
-	int moveX,moveY;
+	float moveX,moveY;
 	qint64 lastMoveUpdateTime;
 
 	struct ActionCacheEntry
